requirements.txt
ultralytics==8.2.103  
opencv-python>=4.8.0
numpy>=1.24.0
shapely>=2.0.4
pyyaml>=6.0.1